+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+* Add `oauthAccessRevocation` to `WebhookNotification`s
+
 ## 3.29.0
 * Add support for `association_filter_id` in `Customer#find`
 * Add support for setting `timeout` and `acceptGzipEncoding` values on construction of `Configuration` instances